pyproject.toml
[tool.poetry]
name = "migrate-nunit"
version = "0.1.0"
description = ""
authors = ["François Perin <f.perin@smartway.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.52.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
